Rewrite all the Javascript
The Javascript code is now moved to a seperate imports file to
maintain ease of overview of the source code, with the aim of
making it easier to maintain in the future as well. Additional minor
features (such as displaying the progress bar upon changing the
architecture and storing the chosen arch/API/variant combination on
supported devices) have also been added.