Simplify references to extraAds[] images
The images are now referenced by simple names, denoting what the
start of the image filename should be. Simple as that - no more
messing around with full paths. Additionally, PNG is set as the
proper enforced default since that is what we want to use for
these kinds of images.