Maintain default behaviour for weird link cases
In case the user presses the shift key or uses the mouse wheel button
to navigate, the browser will want to trigget special behavior rather
than what we have set as the default. In those cases, we should let
the browser do its thing and hope the GA call goes through regardless.