Fix downloads for Microsoft Edge
On Edge, the HtmlCollection can't be iterated over using a for..in
loop. Intead, you got to iterate over the elements using their
indexes. This commit makes it so that these indexes are preferred
when possible and the for..in method is used for when the array
length is unknown (and therefore indexes can't be determined).